pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "skillinquisitor"
version = "0.1.0"
description = "Security scanner for AI agent skills"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
  "pydantic>=2.11,<3",
  "PyYAML>=6.0,<7",
  "typer>=0.16,<1",
]

[project.optional-dependencies]
ml = [
  "huggingface-hub>=0.35,<1",
  "sentencepiece>=0.2,<1",
  "torch>=2.8,<3",
  "transformers>=4.57,<5",
]
llm = [
  "huggingface-hub>=0.35,<1",
  "llama-cpp-python>=0.3.16,<0.4",
]
all = [
  "huggingface-hub>=0.35,<1",
  "llama-cpp-python>=0.3.16,<0.4",
  "sentencepiece>=0.2,<1",
  "torch>=2.8,<3",
  "transformers>=4.57,<5",
]

[project.scripts]
skillinquisitor = "skillinquisitor.cli:main"

[dependency-groups]
dev = [
  "pytest>=8.4,<9",
  "pytest-asyncio>=1.1,<2",
]
